Drop the pure-Python side, scope this to the C extension only
The pure-Python implementation has the identical shared-counter
shape (a module-level _version array bumped by every instance) but
no locking of its own anywhere else, so fixing just this one counter
there doesn't buy real thread-safety and is out of scope here. Move
the regression test out of test_version.py (which parametrizes over
both backends) into test_multidict.py alongside the other
c_extension-marked free-threading tests, using MultiDict/getversion
directly instead of the dual-backend fixtures.
