+++ CHANGES/1444.bugfix.rst
@@ -1,7 +1,7 @@
-Fixed a data race on the free-threaded build where the shared version
-counter used by ``getversion()`` (module state in the C extension, a
-module-level counter in pure Python) was bumped with a plain increment
-instead of an atomic op. The counter is shared by every multidict
-instance, so two threads mutating different instances at the same time
-could step on each other's update and hand out a duplicate (or
-non-monotonic) version number -- by :user:`asvetlov`.
+Fixed a data race on the free-threaded build of the C extension where
+``state->global_version``, the version counter shared by every
+multidict instance and used to derive ``getversion()``, was bumped with
+a plain increment instead of an atomic op. Two threads mutating
+different instances at the same time could step on each other's update
+and hand out a duplicate (or non-monotonic) version number
+-- by :user:`asvetlov`.
